chore: #131 폴더 분리
