+++ Spoony-iOS/Spoony-iOS/Source/Feature/Explore/Model/ChipColorEntity.swift
@@ -1,0 +1,15 @@
+//
+//  ChipColorEntity.swift
+//  Spoony-iOS
+//
+//  Created by 최주리 on 1/22/25.
+//
+
+import Foundation
+
+struct ChipColorEntity {
+    let name: String
+    let iconUrl: String
+    let textColor: String
+    let backgroundColor: String
+}
